+++ .vscode/settings.json
@@ -6,6 +6,7 @@
     "Contáctanos",
     "doloribus",
     "Encontranos",
+    "formik",
     "Formik",
     "Lulo",
     "magnam",
